Updated End of Level Score Calculations
- scores are remembered between sections and added on to
- continues are counted into the score
- a continue restarts the screen, rather than the level

+++ Assets/Scripts/Systems/CutsceneSystem.cs
@@ -15,7 +15,7 @@ public class CutsceneSystem : MonoBehaviour
     private void Awake()
     {
         _gameManager = GetComponent<GameManager>();
-        _gameManager.OnLevelEnd = (int _, float _) =>
+        _gameManager.OnLevelEnd = (int _, int _) =>
         {
             StopAllCoroutines();
         };
